mixer: Let script be called all the time
A mixer script can be used/useful in many cases, e.g. when mute is
called maybe we want to run a script to turn on the amplifier, or send a
notification.

So lets split off the script mixer to be called individually. This means
we should be able to provide the option `--mixer` at least twice.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>